# backend/iisProject/src/main/java/hr/algebra/iisProject/dto/AuthRequestDTO.java
package hr.algebra.iisProject.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class AuthRequestDTO {
    private String username;
    private String password;
}