# backend/iisProject/src/main/java/hr/algebra/iisProject/dto/RefreshTokenRequestDTO.java
package hr.algebra.iisProject.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class RefreshTokenRequestDTO {
    private String token;
}